+++ satel_integra_enh/__init__.py
@@ -1,3 +1,4 @@
 """Top-level package for Satel Integra."""
 
 from .satel_integra import AlarmState, AsyncSatel  # noqa: F401
+from .connection import CircuitBreakerState  # noqa: F401